+++ src/scope/network/exceptions/unexpected-network-error.js
@@ -1,4 +1,15 @@
 /** @flow */
 import AbstractError from '../../../error/abstract-error';
 
-export default class UnexpectedNetworkError extends AbstractError {}
+export default class UnexpectedNetworkError extends AbstractError {
+  message: string;
+  constructor(message: string) {
+    super();
+    this.message = message;
+  }
+  makeAnonymous() {
+    const clone = this.clone();
+    clone.message = this.toHash(clone.message);
+    return clone;
+  }
+}